karm-gfx: Use floats when possible.

+++ src/libs/karm-kira/checkbox.cpp
@@ -30,7 +30,7 @@ struct Checkbox : public Ui::View<Checkbox> {
 
             if (_mouseListener.isPress()) {
                 g.strokeStyle(Gfx::stroke(Ui::ACCENT600).withWidth(1).withAlign(Gfx::INSIDE_ALIGN));
-                g.stroke(bound(), 4);
+                g.stroke(bound().cast<f64>(), 4);
             }
         } else {
             g.fillStyle(_mouseListener.isHover() ? Ui::GRAY600 : Ui::GRAY700);
@@ -38,7 +38,7 @@ struct Checkbox : public Ui::View<Checkbox> {
 
             if (_mouseListener.isPress()) {
                 g.strokeStyle(Gfx::stroke(Ui::GRAY600).withWidth(1).withAlign(Gfx::INSIDE_ALIGN));
-                g.stroke(bound(), 4);
+                g.stroke(bound().cast<f64>(), 4);
             }
         }
